Add scheduler
Adds a minimal scheduling util class that allows periodic and "one-shot"
tasks to be scheduled. The main loop simply invokes "tick()" and
`scheduler` will run any task that has reached its deadline.

This simplifies the main loop quite a bit and makes it possible to
identify the next task to run instead of open-coding time checks.
Hopefully we can use that to sleep when there is nothing to do, saving a
bit of power in the process.

A native_tests platform is added to test the scheduling library in a
regular development environment. I am not sure this part is done in the
best way though.