Add a badge class to manage the global state
The badge class managed the current logical state of the badge.
Currently it mostly hides things that were global, but it will
gradually evolve to represent the "application logic" that drives
the display, leds, and communication components.

Follow-up changes move the rest of the main loop to become task-based
which are ran by the scheduler and scheduled by the badge and its
various subsystems.