Make LED animations and button polling task-based
The LED strip animations are now run as part of a periodic task
(see strip_animator.cpp) with a variable period.

To replicate the existing functionality of increasing/decreasing levels
when the up/down buttons are pressed, a "button watcher" task is added.
This task samples the states of buttons periodically
(every 10ms right now, configurable in config.hpp) to debounce the six
buttons (up, down, left, right, ok, cancel).

The "raw" button readings are converted into "button up", "button down",
and "button down repeat" events that other components can listen for.

These two components are grouped under a "badge" class. This class
represents the logical state of the badge; currently that means:
  - current animation,
  - state of the buttons,
  - the user's "social level".

In terms of code size, almost all of the increase comes from the use of
Serial.printf() which I left to make it easy to test the buttons for the
moment.

Another contributor (~200 bytes per task) is the fact that the scheduler
is entirely implemented in its header which causes multiple copies of
the "schedule" method to be replicated (one per task). I'll move this to
its own compile unit soon, but I don't think its much of a problem for
the moment.