Add a menu screen
The menu screen is a menu template that can be updated with various set
of choices. This allows sub-menus to be implemented by updating the
current choices of the menu.

Currently, the badge sets a place holder set of choices to demonstrate
the rendering features and the overall look and feel. The actions for
the various choices consist in logging the name of the choice.

To try it, press 'Cancel' to exit the 'idle' screen.

The place holder demoes the following features:
  - scrolling,
  - drawing a "scrollbar" when needed,
  - highlight current selection,
  - action executed when pressing 'Ok' or the right arrow,
  - left arrow goes "back" to the idle screen.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@gmail.com>