Add a string_property_editor class to edit the user's name
The string property editor is used to allow the user to edit any string.
For the moment, it is used to edit the user's name. It allows the
characters: [A-Z], [a-z], ' ', and '-'.

The up/down arrows allow you to change the current letter (rendered with
negative colors) and the left/right arrows allow you to change the
current letter.

The 'X' button can be used to delete characters. To leave the menu, hold
the 'okay' (✅) button.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@gmail.com>