Add an nsec logo for the splash screen
The logo was scaled down in gimp, exported as a C file using Gimp's
export facilites. Then, an hand-written util converted it to a format
suitable for use with the badge's screen.

The assets and generated code are included under the logo folder. I'm
not sure if we want to keep them around, but they can be useful for the
moment.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@gmail.com>