Network: refactor into a proper protocol state machine
Refactoring the network code into a proper state machine will make it
easier to implement the retransmission logic in a follow-up commit.

Retransmits are sometimes necessary since other tasks (the LED strip
animation task, for one) periodically disable the interrupts which
causes the SoftwareSerial to miss the beginning of some transmissions.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@gmail.com>